+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Bugfixes
+
+* Add `for`/`id` attributes to filter labels and inputs for accessibility (#91)
+
 ## v0.9.0-beta.4 (2026-01-26)
 
 ### Features